requirements.txt
aiogram
aiosqlite
sqlalchemy[asyncio]
loguru
mss
plyer
PySide6
PILLOW
loguru
winsdk
pynput